Initial commit..
In order to allow users to dynamically edit their selection ( of students
and/or questions ), it is important to have some scheme for storing a
snapshot at any given time. XML seems to be an obvious choice.

The first script - db2xml - mimics DB -> XML conversion. Right now, we
don't have a DB.

The second script - xml2Tex - probably wouldn't change much. And it is
this script that does the heavy lifting of iterating over XML and
sequencing the concatenations of .tex files into one large document
( that is then compiled into PDF ). It will also generate student specific
QR codes