DB -> XML: Perform intermediate transformations here itself..
As far as I can tell now, one would first get records from the DB
(representing students and questions) and some/all information
in these records would be written into an XML file.

What is written into the XML file will not be a simple decompilation
of the DB record. Some intermediate processing might be required
and/or preferred. For example, we might want to maintain the first
and last names as separate fields in the DB - but in the XML we might
want to merge them into one attribute called 'name'