DB->XML : While you can, generate intermediate QR codes..
The generated code is the bit that comes from the question and will
hence not change. However, the final QR codes will also incorporate
student ID and would hence be different for each student - even for
the same question.

Also, in merge_attributes, provide an empty 'merging scheme' for
DB fields that need to be transferred untouched to the XML. See
explanatory comments in the code