Added code to generate XML for question paper...
The final document contains updated QR codes for each question
for each student.

This XML will eventually be generated from with the Rails code itself -
right after we get the list of students and selected questions from the DB.
But for now, we are using the proxy XML files - students.xml & questions.xml -
as inputs for generating this file.