Defined steps for .TeX -> PDF conversion...
Note that this makefile is *not* expected to be called manually
from the command line. It is expected to be called from within
the Rails app - right after student and question selection has been closed.

However, if you do invoke it at the command line,
here are some things to know:

1. Usage: make session=<session-id>.
2. 'question_paper.xml' needs to be present before invocation.
   Which therefore means that one might also have to do the whole
   {questions,students}.{db -> xml} -> question_paper.xml conversion
   oneself