Upgraded to Rails-3.1.1 ( from 3.0.4 )
RJS has also been extracted out of 3.1.1. However, reference to it
in development.rb remained - leading to error messages. Therefore
commented it out.

A warning message for an already defined constant still appears.
This is a known issue when upgrading to 3.1.1. The solution is
to either downgrade to rack-1.3.3 ( from the current 1.3.4 ) or
to wait for 1.3.5. We will wait