Bare shell added for Parent model..
Right now, it has only one boolean field for whether or not the parent
is the mom. Might be used to track which parent takes more interest
in their child's education...

Still not sure about whether or not to ask for personal information
from the parent. How could one use it?