[New Model] : DBQuestion
This model captures those aspects of a question that are independent
of how and where it is actually used ( in quizzes ). Hence, things
like path to TeX code, the number of times this question has been attempted,
what topics it touches upon etc. etc. are part of this model

All other information - like the questions position in a quiz or
whether or not its a teacher's favourite - should be captured
someplace else ( specifically, in the Question model )