Added new phony target - jpeg - to generate JPEG preview from PS...
Currently, jpeg is the default target. It generates a preview of
the answer key. This way, teachers can not only see the question
but also the answer before deciding to select it for their quiz.

JPEGs are being generated because they might be easier to load
using javascript libraries. However, if loading PDFs isn't that much
different or any more complicated, then we might just remove this
jpeg target all together