Initial commit of code to provision a DB slot for a new question...
The person adding a new question would not know what the last question
in the DB is. Hence, what is needed is a way for the system to
block a slot in the DB for the new question. Blocking a slot means
not just adding a new entry into the DB, but also creating a new folder
and adding that new folder into the Git repo