[annotate]: Fixed ArgumentError problem..
The problem was arising due to incorrect calls to :belongs_to.
I was assuming that X :belongs_to A,B,C is equivalent to
X :belongs_to A, X :belongs_to B and X :belongs_to C. Not true.

This was a basic syntax error. However, it took me a long time to fix