Initial commit of WelcomeController..
Here is how the application works. The root in routes.rb
is set to WelcomeController#index.

The 'index' controller renders welcome/index.html.erb ( not yet haml )
if no one has logged in. See hack in welcome_helper.rb that allows
form creation even when current_account.nil?

When the user logs in, the app redirects back to root - and therefore
to Welcome#index. But this time current_account is not nil and the
application redirects to another controller - depending on what
current_account.role is