Rationalized Question model further...
Question properties - like whether or not its an MCQ or has multiple parts -
should be captured in the DB. And it is also from the DB that these
values should be picked up when rendering views.

The QR code can thus be simpler. If it can provide the question ID,
then all the rest can be pulled from the DB