Added difficulty column to Questions & Syllabi..
The syllabi table maps courses to topics. However, the difficulty
level of topics is also an important characteristic of the course.
In the same course, some topics might get basic coverage while
others get advanced treatment.

The difficulty of a topic in a syllabus is then matched
with the difficulty of questions on that topic in the DB.

There are 3 difficulty levels : basic, intermediate and advanced.
A question can have only one of these levels. In the syllabus,
difficulty could be some OR combination of the above