Added BroadTopic model..
A broad_topic :has_many :specific_topics and a
specific_topic :belongs_to :broad_topic

Note that there is no connection between any of the topic models
and Subject. Why ? Because its possible that a topic might be
covered as part of a subject which one would not ordinarily consider
a topic to be part of.

For example, in the UK, mechanics is part of maths whereas everywhere
else it is strictly part of physics. Or for that matter, VAT calculation
is part of ICSE maths whereas it is probably covered in Accounting in CBSE.