Create button-sets for "resources" with multiple associated buttons...
Note the use of the term "resources" and not "forms". A "resource"
- as I intend it - is a logical grouping of HTML elements that provide
some information to the user and allow certain actions to be performed
on the underlying DB via it. A formtastic form made based on an ActiveRecord
model is one example.

Anyways, I think that in order to keep the HTMLs clean and self-consistent,
its best to hand-code "submit" buttons ( rather than use f.buttons provided
by formtastic ). Formtastic is great. But sometimes it creates too deeply
nested a structure