[GradeDescription] : See no need for views for create and update actions..
In the JS and AJAX centric world we are trying to create, the user should
have little need to navigate away from his account console. This therefore
means that for models like GradeDescription, all that we need to do is render
pop-up forms for the user to fill in. That can be done with partials.
And the controllers should either render just the HTTP header or some
minimal JSON data