Added controller for Grades...
Increasingly, it seems to me that models like Grade, Quiz etc. should
have controllers with only 2-3 actions max. The other option is to implement
actions that would otherwise be in models like Grade in controllers of
"human" models - like Admin, Teacher etc. But that would require nested
routes - which I am not too keen on