Added 'rabl' gem to the list..
This is a really useful gem. It allows (easy) creation of JSON responses
in much the same way that one creates HTML responses (using erb / haml ).
Given that I expect much use of JS, AJAX and JSON, incorporating the gem
seems like an obvious thing to do

The alternative - really, the standard method - is too complicated
and gets really confusing when deep nesting of models is involved. Moreover,
as the gem's author says, the controller should not have much/any code
related to rendering a response - be it HTML , XML or JSON. Responses
should be defined strictly a view file. I kinda agree...