Incorporating a new attribute - marker - for input and link elements...
The plan is to have this attribute on all HTML elements the user can interact
with. Mostly, this means <input>, <a> and the like...

No assumptions are made for what value a marker should have. But the plan
is to give markers value that would either:
  1. help decide whether the containing element is the one we want, OR
  2. provide critical input for further action

I decided against using standard HTML attributes because as things stand now:
  1. class attributes are too non-specific and are anyways used exclusively
     for defining CSS styling
  2. id attributes are too mangled to be worked with easily - especially
     when generated by things like formtastic