Separated JS asset for admin from that for the overall application..
This is the way it should be. Once you log in as a particular role,
only the JS - and even the css - for that particular role should load.

This requires creating separate assets for each role. Hence, admin.js -
which was previously just another .js file - is now a manifest file.

'application.js' is effectively empty for now and application layout
simply yields. As far as I can tell, only WelcomeController would
default to using application layout and assets. All other roles will
 - nay, should - have their own assets