Radio-buttons and checkboxes can now be part of a table row...
In HAML, add either {:notitle => :radio} or {:notitle => :checkbox} to
the list of columns. No radio-buttons or checkboxes will be created
because of this. But the table heading will get an empty space below
which they would eventually be lined

In the JS, add something like [null,'radio | checkbox', <optional marker>]
to the array of arrays that is passed to 'createTableRow'. That's it..

However, the most important fix in this commit is probably in the CSS.
Know that if table alignment is screwed up, it might not be because you
got widths wrong. It might be because you got the heights wrong