Created a toolbox folder within view..
The idea is to house small, reusable bits of HAML/HTML in a toolbox so that
these can then be picked up later to stitch together an interface. For example,
one might want to implement a search box with 2-3 inputs. Simply pick the
bits and bobs from here and get going ...