Summary tables given an additional class sttribute - summary-table...
This class attribute is required to bind ajaxSuccess() to all summary tables.

Thing is that there can be multiple events that eventually need to
update a summary-table. And these events are bound to different DOM elements.
And so, rather than tie the updation logic to the DOM element, it makes
sense to tie the logic to the summary table itself.