File wherein one over-rides element attributes...
The first use case is over-riding the action attribute of forms
generated by formtastic. Sometimes, the value you need is not the
value you get post-creation. That's when one overrides the value
in this file

I suspect this file would stay small. But a version of it should
exist for each role