Added search functionality for courses ...
Note to self : Change grade column in Course to klass.
Grade is already a model with totally different connotations.
And we can't use the word 'class' in the way its normally understood in
schools because 'class' is a reserved word in many programming languages.
Hence, any code using it will either not run or behave funny