Moved files from topics/ to specific_topics/ to stay consistent..
Thus far, I have followed a scheme in which if there is a model with a
controller, then all view fiels associated with it are stored in a similarly
named folder ( as generated by rails generate controller ... )

The above was an exception I intentionally introduced because the name topics/
seemed more apt than the underlying model name. Reverting that change. I guess
its more important to be consistent in one's code organizing scheme than
to like any names