[NeXT Step] : Initial framework for editing sylllabi added ...
The framework includes a "mega" table in which specific topics are listed
under broad topics. A specific topic can be selected using the checkbox
alongside it. And for the course in question, the target difficulty level
can also be set for that specific topic

The plan is that instead of doing one big form submission, ping the
server each time a checkbox is checked / unchecked OR the difficulty
level is changed.

As it is, defining syllabi will never be an everyday task. At most,
once a year