New option - prefix - for dropdown_menu_for helper method ...
The prefix option can be used to modify the 'name' attribute of the
final <select> element this method creates.

It is the name attribute that is used in defining the params hash
when submitting forms. So, a form element with 'name' = 'ceo[apple]'
and value = 'steve jobs' would be submitted as :
   params => { 'ceo' => {'apple' => 'steve jobs'} .... }