Can't do without forms !!!
My earlier assertion about being able to edit syllabi without forms
(using quick pings to the server) would not hold - or at least would
be too messy to implement. Hence, enclosing #edit-syllabi-megatable
contents within a form

There is a hidden input field that would be filled - in JS - with the
course_id. That, and the user selection of specific topics, is what
we need to create / update a syllabus