Added a new class attribute - empty-on-putback ..
Apply this attribute on any DOM element that needs to be
emptied before being put back into the "toolbox"

Previously, this action was performed on '.data' w/ the
added restriction that only the first .data would be emptied. This
is unnecessarily restrictive. Moreover, .data is not a good choice
of attribute name for elements that need to be emptied

Associated changes in JS to follow