Replaced the one 'top_panel_link' helper method with 2 helper methods ..
The 2 new methods - main_link & minor_link - are the ones to use in view files.

Usage : [main_link | minor_link] :for => X, :href => Y,
           :with => { [:panels => {[:side => A], [:middle => B], [:right => C], [:wide => D]}],
		              [:controls => G] }

Why was this done ?
-------------------
The old helper method assumed that other than 'main-links', no other type of link
would ever need to load a panel or controls.

But, in the new interface, every link is 'dynamic' in that sense
( barring links like 'login/logout' )

The other commits are of files that are impacted by the change in helper methods