First steps towards allowing easy edit of faculty rosters ..
What makes editing faculty rosters different from other edits thus far
is the fact that what needs to be changed depends on which school and
which teacher is selected previously. Also, when editing rosters, one can't
stop at showing just the study-groups a teacher is teaching now. One has
to show all study-groups so that additions and deletions can be made