Massively Imp : Defined 10 possible yardsticks as per new scheme ..
The old yardstick scale was still not defined sharply enough. A little
more thought, and we realized that every student response will have the following
3 characteristics :
  1. whether or not it captures all the required critical insights
  2. whether those insights have been formulated into equations
  3. whether those formulations have been 'calculated' to arrive at an answer

Each of these characteristics can be either absent, partially present or
fully present in a response.

The above scheme leads to 10 well-defined yardsticks (no, not 3^3 = 27)
As these are known, we can simply render the scale at render time