First steps towards transitioning to PostgreSQL ..
This is the industrial strength we would need to use. And Heroku recommends
that PostgreSQL be used even in the development environment

Ref : http://www.mcbsys.com/techblog/2011/10/set-up-postgresql-for-rails-3-1/

You will need to install Postgres locally. However, it requires
username and password. Which is database.yml is being removed from
the repository ( but not the project. You will need a local copy )