Incorporating CoffeeScript ..
Plan is to convert all JS into CoffeeScript. Why now?

The JS code needs to be re-factored. It is responsible for all
interactions on the site and the code has gotten kinda messy.

So, if its going to be re-factored, then I want to write the
least new code I can when doing it. That's where CoffeeScript comes in