No need for a Trojan horse in new-student form ...
inserting a trojan horse and editing the form action just before submit
are 2 ways of customizing a form for the currently selected record

but when it comes to code simplicity and consistency, the latter
seems better (even though the former has a cooler name). Hence, am
going to stick with just one approach - changing the action attribute.
Bye, bye Greeks !!