[Weird] : For teachers & students, ensure non-offensive timestamp in login
For the above 2, we use convert the time of creation into a base-36 string.
As its base-36, the whole english alphabet is used. And it can happen that
the 3-4 character word that results is an offensive word.

For example, we wouldn't want a login like rsachdeva.dick or
rsawant.tits, would we ?