Converting #new-examiner to a pop-up form ..
If an activity will be done frequently, then it needs to be done
quickly. In that case, it makes sense to render any form or table
in a panel

However, if an activity will be done infrequently, then using the panels
to effect it is bad usage of the available real-estate. In such cases,
explore the possibility of using pop-ups