First steps towards deploying jQuery sortable lists ...
For lists that could potentially get very long - like those for
specific topics - one needs an elegant way of presenting the
individual list items to ease comprehension. jQuery sortable lists
is the option I am going to try and deploy now

This would usually entail showing multiple list items part of
a smaller group - as opposed to being part of just one large group

Sometimes, this grouping maps with what is already in the DB - like
every specific topic belongs to a broad topic. And sometimes, this
grouping is quite context-specific and hence needs to be reflected
in the view - like the difficulty levels for a topic in a given course