[NeXT] : First pass implementation of a new way of presenting long, hierarchical lists..
The list of macro & micro topics are the kind that this new way hopes to tackle.
Not only does each list have many items, each micro-topic also belongs to a
macro-topic

The solution is to present each separately - macro in the side(middle) panel
and micro, therefore, in middle(right) panels

Micro presentation still to be done. But it seems that :
  1. The need for #wide-panel would go away
  2. Our inability to select a micro-topic for editing/deleting in the old scheme
     would also be addressed in the new scheme