Have had to change the params key used to get course from :course_id to just :course
This has sth. to do with how the current JS code updates radio-button
URLs through reg-exp matching. It basically looks for any id= and changes
it to id=<something>, where <something> = marker attribute on the radio-button

But if the URL is of the form a/b.json?id=<x>&course_id=<y> - for example -
then even though id != course_id, it will be because of the above
substitution logic

Lets just put this down as KPNS for now