This is the God element - kinda like the God particle CERN researchers...
.. are searching for. It has everything in it - radio-buttons, check-boxes,
a label, <select> for difficulty etc.

In cases where the list to display is already known, iterate over the DB
records and render this element. Then, in JS, disable & hide the bits you don't
want to show in the list. So, if you want to show checkbox instead of a
radio-button, hide the latter but not the former