Setting default marker on control and side panels ...
Consciously not setting it on middle, right or wide panels. These panels
should get their marker from any clicks within them - and only from those clicks
The side panel - as it can be anything - is not guaranteed to have a non-null
marker because its not guaranteed to have radio buttons within it

This change is being done now because the way our form work, they use
the marker on the panel within which they are. An undefined marker
only leads to params{ id => 'undefined' }, which is really annoying