Cleaning up logic pertaining to what AJAX request to send on radio-button click ..
The old scheme of setting a 'url' attribute on every radio-button was unwieldy,
ugly, unintelligible and unscaleable.

The new scheme is similar to the kind of logic implemented in receiver.js.coffee.
The decision of which AJAX request to send is now done once the click
event has been captured by some pre-delegated parent DOM element

Also, as we now don't embed an attribute on every radio-button,
the HTML is cleaner. Plus, less revealing to anyone who does
'View Source' in a browser.

Gaps to be filled-in in subsequent commits