Another round of code refactoring ..
The underlying logic governing macro/micro topic interplay was not
very elegant. And this commit represents the first steps towards
addressing that. Basically, a macro or micro topic entry can now
lie within only 1 of 3 lists : the master, selected or deselected.
If one needs to display a macro or micro topic, then one moves the
whole list to the desired insertion point

In the process, I have realized that creating some sort of namespace
for functions would be fruitful for the future. So, some of that is also here