Change the login username and email when profile information is changed ..
Currently, this is done for teachers and students only. The idea
is that if a spelling mistake is corrected in the first or last names,
then the login should reflect the corrected spellings. Hence, change
the login and email everytime there is a save of teacher or student record

Note that updation of student & teacher profiles would NOT be a frequent
operation. And the overhead, therefore, of doing the above is acceptable