Added a new method - currIndex - to return the index of the currently previewed image
In order to select a question for a quiz, we first need to know
what its position is in the list of options being previewed. That is
what this function provides

The 2nd piece of the puzzle would be the DB id for that question.
This information would be stored as attribute 'marker' on the
<li> within .ppy-imglist