Trimmed Quiz model ..
Quizzes, once made, would most probably be re-used again by the teacher.
Hence, the model should capture only that information  which will remain
invariant in each use

The use-case, as far as we can tell now, would be that the teacher would
build a Quiz once and then "apply" it at a later time to a group of students.
It is then that slots for responses would be created in the DB and
QR codes generated