[Bug] Don't touch the #control-panel on radio-button click ..
The control panel gets its marker attribute during the initial render.
And the marker is set equal to the loggable_id - NOT account ID - of
whosoever is logging in. Thereafter, this marker should not be touched.
Its provided as a convenience for when one needs to send ?id= with
an AJAX request