Remove - not purge - ppy-placeholder on call to preview.loadJson
Popeye() is not re-entrant. In other words, changing the img-list
is not enough to refresh the display completely and correctly.

What is needed is a fresh call to popeye() with the new image list.
But one also can't call popeye() on an existing .ppy-placeholder without
messing the display up

The only way, therefore, is to do everything from first principle.
Hence, clone the virgin document-preview element, fill it up with
the img-list and then call popeye() on it