A panel should get a 'marker' ONLY if a radio-button within it is clicked ..
The old way in which a panel 'marker' attributes could percolate from
side -> right was messy and wrong. For one, a panel could have a non-null
marker attribute without anything being selected within it. And that might
just be plain wrong

This commit would break some existing behaviour in the Admin console.
But the new scheme is right and hence the breakages will be fixed