GradedResponse now :belongs_to a QSelection
The QSelection represents a particular question, picked for a particular
Quiz, laid out on a particular page. All students receiving the said Quiz would
therefore get - and give responses to - the same questions on the same page

There is therefore no need to maintain the Quiz, Question or page number
information in the GradedResponse model. All these can be got from the
QSelection