[Popeye] First steps towards img-caption persistence ...
Popeye collects all the captions at load time and stores them in a local array
variable. Thereafter, it uses this array to pick the caption to show

However, if one wants to change captions after loading - based perhaps
on events like key-press - then the above caption storing scheme does
not cooperate. It is this that I have tried to fix in this commit