[Bug] : Successive clicks on 'make new quiz' link would create multiple previews ..
As long as #document-preview was listed as a #wide-panel requirement for the above link,
resetPanel in the core would place #document-preview within wide-panel.

Thereafter, popeye would execute on this newly inserted document-preview -
even if one existed before. This leads to the display going hay-wire

The cloning & insertion of document-preview from blueprint was being done by
 #btn-candidate-questions and should be left like that