[Admin] Upto 5 students can be added at a time now ..
.. And, instead of writing their first & last names, simply
write their full name

Simultaneous assignment to study-groups to follow.

Note, however, that with this chainge, the timestamp appended to each
new student's username *can be* the same for multiple students.
This will happen if the student records were created within the same second

However, the assumption is that their names would be different enough
not lead to a conflict