Defined method in Quiz to issue Savon request for building TeX
.. In the controller, the create action renders the JSON for the
response and sends it to the client. And if the response was an
error, then it destroys the just-added record for the Quiz in the DB

would have loved to do this using after_create. But call-backs cannot
return anything more than a true or a false to their calling  operation
(like save or create)