If a testpaper is destroyed, then remove associated data about responses & course-pack ..
A testpaper is an instantiation of a Quiz and is never really destroyed on its
own. Its only destroyed if the parent Quiz is destroyed. In which case, it makes
sense to remove all data pertaining to grades earned on that Quiz and removing
the Quiz from a student's course-pack