Destroy all graded-responses associated with a qSelection if the latter is destroyed ...
If a QSelection - that is, a Question <-> Quiz mapping - is being destroyed, then
it means that some other destroy action - most probably on the parent Quiz -
is being called. And so, if the Quiz is being destroyed, then there is little
point in retaining the responses to that Quiz

As noted in an earlier commit, however, destroying a Quiz should never be
a user-initiated action. It should be done by a cron-job