Overhauled interface for #yardsticks-summary ..
We are now rendering a vertical jQuery tab-panel that looks like those
colour-charts one uses to pick colour for one's walls. Much more intuitive

The above is done in the #side-panel. The remaining 2/3 space will be
taken up by the #wide-panel to show #document-preview

the preview would contain example responses of varying degrees of "quality"
- the kind captured by our new colour-chart