First steps towards changed interface for grading ..
We now need a <canvas> element in which to edit the JPEG and a document-preview in
which to see the solution. Not done in this commit - but that's what we need

Moreover, we don't need the examiner to dive from quiz to student to pages etc. etc.
Just show them the pending scans and some sort of a control-panel for grading
Any intelligence on what to show and how should be built into the model/controller