Reconciled inconsistencies when creating a school ..
Only name, tag, zip-code and email are required. This is reflected
now in both the model and JS-validations

The above 4 are required to create not just the model object
but also the associated Account object. We were seeing a problem
whereby a school was getting created but not its associated account -
leading to a crash when schools#show was called