Track radio-button clicks in a flipchart on the containing tab itself
Previously, we assumed that a radio button in the first tab was special
and only its marker needed to be tracked. This is no longer true. We should
track the marker for any radio button in any tab.

Where should we track it then? On the containing tab! Where else?