Reduced minimum password length requirement from 6 to 3
The default password on account creation is th zip code. And its length
changes from country. In the english speaking world, we have
   1. US: 5 or 9
   2. UK: 5
   3. Australia: 4
Just to be safe, we are going with 3

If the password length validation fails, then an account is not  created for the new
record. And then everything else falls to pieces