Multipart question layout defined at the time of tagging ..
In the absence of \newpage directives, LaTeX would lay out all the questions
on the same page - leaving space for solution in proportions defined by
the \mcq, \halfpage & \fullpage commands.

The result would be undesirable most of the time. Hence, the # of subparts
on a given page and the page breaks need to be defined upfront