Multi-part support added to preview::loadJson ...
With multi-part questions in play, one has to prepare for questions
that can span multiple pages.

Not only does this impact how popeye's image list is populated, it also
impacts how it is traversed. In particular, 'back' and 'next' operations
have to be redefined to jump as many scans as needed to get to the previous
(or next) question