Renamed 'attempts' column to 'n_picked' in Question ..
The new column will be used to track how many times a question has been
picked to be included in a Quiz.

The count can be used as the basis of an incentive scheme whereby question
writers are rewarded for how many times questions written by them are picked.
Much better than rewarding them for the # of questions written