Fixed issue of misbehaving accordion within #past-quizzes ..
Basically, the problem was because two JS methods - .tabs() and .accordion() -
were being called on the same element - #past-quizzes. This is really not allowed

The solution is to render #past-quizzes as a tab() and an element within it
as accordion()

While at it, I DRY'ed up the code by removing nail-file and katana. We already have scroll and
that should be enough