Added <a> to swiss-knife. Also, modified customize method ..
Previously, the label within the swiss-knife was always retained.
But sometimes, and element - like an <a> - can not only serve the
purpose a label does, but provide some other functionality too (like download)
On such occassions, it makes little sense to retain the label. Moreover,
treating a label as differently from others just breaks code symmetry