First steps towards getting rid of faculty rosters ..
Here is what we are drifting towards:

What klass a student belongs to is a fundamental property of the student. Similarly,
what klass a sektion/grouping is for is its fundamental property. A student
can therefore belong to any sektion that is for the klass the student is in

Control over defining/editing sektions is now with the teacher. She can create
'public' sektions - visible to every other teacher in school - or private ones
- visible only to her

Faculty rosters is therefore a redundant concept. A teacher needs to see only
sektions of students - grouped by the klass/grade level - even if they haven't
been defined by her