generate_suggestion_form added to gutenberg.yml ..
Important if one wants the HTTP header in a Savon request to be exactly
as one wants it. Things seem to work though even without the perfect HTTP
header - as was happening till now. But the question is - just because
one can get away with it, should one get away with it?