Allow form submissions via PUT or POST ..
Forms where we had explictly specified method => :put had started to behave
weird. The first time around they worked fine. But the second time around,
we would get a "route not found" error because _method in the HTTP headers
was no longer = 'put'

Given that a form can be submitted only as a POST or a GET (not preferred),
it seems only logical to allow POST to map to a route that was previosuly
exclusively PUT