Calibrations can now be enabled (default) or disabled ..
If a calibration is disabled, then a teacher will not be able to see it
Moreover, yardsticks/logical_next will not find it and hence the
underlying yardsticks would be disabled for the examiner also

However, a calibration can *** never *** ever be deleted. There might
be grades & marks associated with it that have been assigned to responses.
If we deleted the calibration, then the DB would go whacko