Standardized JSON to use ticker when ticker required + removed colour coding of marks ..
We don't want to suggest that green is better than orange is better than pink
when showing numerical marks. Hence, they are all shown in a generic gray box.
In the student console, we will go a step further and hide the marks altogether.
A student will have to click to make the marks temprarily visible