[bug]: Ajax request not getting issued when clicking on fdb-panel tabs
The shared panel is non-empty - and will stay non-empty. And when
a tab-pane is non-empty, ajax requests are not issued if tab has
nopurge-on-show

Hence, added a class 'static' - for static content - to fdb-panel.
Allows the ajax request to go through. But overall, not a happy
situation. Need to think more deeply about the minimal set of
class attributes to set on nodes and how the attributes impact purging
- none, recursive, shallow etc.