Bypassing BuildQuiz throughout ..
At best, its a wrapper over CompileQuiz.

Moreover, we in order to show the just queued quiz, we need
to save it in model / controller methods with the correct job ID.

If its saved in BuildQuiz, then the job ID wouldn't be stored until
the perform method is called. In which case, we would have to wait
to get the time estimate