Re-incorporating client_side_validations into the app
Not only is this a better way to do validations, it also allows us to
auto-login users on registration

Previously, our forms used AJAX and validations were server side (returning a JSON).
But to auto-login, we need to call redirect_to. Its not possible to call both
redirect_to and render :json together. Hence, the need to separate the
two - validations and auto-login