[#75]: Allow teachers to add students whilst allowing students to register seprately
Teachers can create a sektion and populate it with students by
providing just their names. But we do not require spreadsheet upload - as
the issue required

Students, however, can register seprately and enter the sektion's UID.
If they don't , then they wouldn't see their graded assignments. The choice is now theirs

But if they do, then the account created by them during registration is merged
with the account created when the teacher added the student's name

Please note that this is only the initial commit. Bug fixes should be
expected and will follow