Passing an optional array of question indices to force quiz layout
'qids' is an array of question_ids. If its NOT blank, then it means
that the passed questions have to be laid out in the order given
in the array. Do NOT count on there being too many sanity checks on
qids in the code below

Passing qids is highly NOT recommended. Do this only if you know
what you are doing