All mailing now done by delayed job
Unfortunately, had to undo the custom job definition - DispatchMail -
from the last commit. It wouldn't work because of the way Rails 3 mailer works

Pity, because it was so elegant. But the underlying issue has something
to do with serializing an object versus serializing a class (which is
what is happening now )