Allow paginator enable to respond to click on a .line
By default, paginator is updated in response to every AJAX call made via a tab-click.
Subsequent clicks on a .line within a .tab-content do not usually lead to loading
of paginable data

However, in some cases, a click on a .line can yield paginable data ( like when loading
proficiency data for students ). In that case, it makes more sense to tie the paginator
to the ajax triggered by the .line