Allow resetting state of any tabs rendered within a tab-pane
For example, the students whose names should be disabled (because
they have already taken a quiz) changes with each quiz

However, this was not happening because the ajax on the first left-tab
(representing the first sektion) was not being triggered. And its entirely
possible that the first tab is also the only tab

The ajax was not being triggered because the first tab had class .active
set on it (by bootstrap) from before. Any bootstrap.js probably (and rightly)
does nothing if an already clicked tab is clicked again

Hence, the solution is to allow resetting => unclicking of child tabs
before the first tab is re-clicked