Allow a criteria to be de-activated ...
A criterion - once added - should not be removed. If it is, then
there would be no way of deconstructing correctly which criterion were
chosen for a given feedback if the criterion has since been removed.

Instead, just deactivate the criterion in that rubric for future times.

Downside - We would limit a rubric to 32 criterion - including inactive ones