Flipped loogic for doing student merge
Rather than update the shell account, update the Attempt, Worksheet and
StudentRoster table with the non-shell student record.

In effect, we are doing what we did before. Attaching the current_account
to the shell account was plain wrong because we would then have had a
situation where more than one shell account map to the same loggable